# optimization: put rfn_sum on cuda and do .item() call out of for loop