@turboderp 2024-05-15
I very much appreciate this.

As I recall there's a whole bunch of synchronization points there, as the converter is swapping stuff to system memory like crazy. I think this change likely just shifts those points around a little bit, which would show up in a CPU profiler but may not affect wall time at the end of the day.

The change looks fine, though, so if it does help I'm happy to merge it.