@kallewoof 2024-05-15
> As I recall there's a whole bunch of synchronization points there, as the converter is swapping stuff to system memory like crazy. I think this change likely just shifts those points around a little bit, which would show up in a CPU profiler but may not affect wall time at the end of the day.

Ultimately it should cause a speed increase as it simply does less stuff (calling .item() once instead of repeatedly).

> The change looks fine, though, so if it does help I'm happy to merge it.

I am looking closer at the .cuda() bump and I think I have a few more optimizations for you.

For example, the `test_error` function calls `.cuda()` twice for `x` and `xref`, but I think we can drop the `xref` call completely and instead move the results in `xtest` back to CPU. That way we circumvent moving the target_states to cuda completely. Testing that change now. Should have results in a bit.