@kallewoof 2024-05-16
Closing in favor of #456 but will reopen if that PR turns out to not work out.